@@ -17,14 +17,12 @@ various services to perform the actual business logic operations on the
 underlying data.
 """
 
+# Standard
 import json
 import logging
 from typing import Any, Dict, List, Union
 
-from fastapi import APIRouter, Depends, HTTPException, Request
-from fastapi.responses import HTMLResponse, JSONResponse, RedirectResponse
-from sqlalchemy.orm import Session
-
+# First-Party
 from mcpgateway.config import settings
 from mcpgateway.db import get_db
 from mcpgateway.schemas import (
@@ -48,7 +46,7 @@ from mcpgateway.schemas import (
     ToolRead,
     ToolUpdate,
 )
-from mcpgateway.services.gateway_service import GatewayService
+from mcpgateway.services.gateway_service import GatewayConnectionError, GatewayService
 from mcpgateway.services.prompt_service import PromptService
 from mcpgateway.services.resource_service import ResourceService
 from mcpgateway.services.root_service import RootService
@@ -60,6 +58,11 @@ from mcpgateway.services.tool_service import (
 )
 from mcpgateway.utils.create_jwt_token import get_jwt_token
 from mcpgateway.utils.verify_credentials import require_auth, require_basic_auth
+
+# Third-Party
+from fastapi import APIRouter, Depends, HTTPException, Request
+from fastapi.responses import HTMLResponse, JSONResponse, RedirectResponse
+from sqlalchemy.orm import Session
 
 # Initialize services
 server_service = ServerService()
@@ -102,12 +105,12 @@ async def admin_list_servers(
 
 
 @admin_router.get("/servers/{server_id}", response_model=ServerRead)
-async def admin_get_server(server_id: int, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> ServerRead:
+async def admin_get_server(server_id: str, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> ServerRead:
     """
     Retrieve server details for the admin UI.
 
     Args:
-        server_id (int): The ID of the server to retrieve.
+        server_id (str): The ID of the server to retrieve.
         db (Session): The database session dependency.
         user (str): The authenticated user dependency.
 
@@ -120,7 +123,7 @@ async def admin_get_server(server_id: int, db: Session = Depends(get_db), user: 
     try:
         logger.debug(f"User {user} requested details for server ID {server_id}")
         server = await server_service.get_server(db, server_id)
-        return server.dict(by_alias=True)
+        return server.model_dump(by_alias=True)
     except ServerNotFoundError as e:
         raise HTTPException(status_code=404, detail=str(e))
 
@@ -154,10 +157,10 @@ async def admin_add_server(request: Request, db: Session = Depends(get_db), user
     try:
         logger.debug(f"User {user} is adding a new server with name: {form['name']}")
         server = ServerCreate(
-            name=form["name"],
+            name=form.get("name"),
             description=form.get("description"),
             icon=form.get("icon"),
-            associated_tools=form.get("associatedTools"),
+            associated_tools=",".join(form.getlist("associatedTools")),
             associated_resources=form.get("associatedResources"),
             associated_prompts=form.get("associatedPrompts"),
         )
@@ -174,7 +177,7 @@ async def admin_add_server(request: Request, db: Session = Depends(get_db), user
 
 @admin_router.post("/servers/{server_id}/edit")
 async def admin_edit_server(
-    server_id: int,
+    server_id: str,
     request: Request,
     db: Session = Depends(get_db),
     user: str = Depends(require_auth),
@@ -195,7 +198,7 @@ async def admin_edit_server(
       - associatedPrompts (optional, comma-separated): Updated list of prompts associated with this server
 
     Args:
-        server_id (int): The ID of the server to edit
+        server_id (str): The ID of the server to edit
         request (Request): FastAPI request containing form data
         db (Session): Database session dependency
         user (str): Authenticated user dependency
@@ -210,7 +213,7 @@ async def admin_edit_server(
             name=form.get("name"),
             description=form.get("description"),
             icon=form.get("icon"),
-            associated_tools=form.get("associatedTools"),
+            associated_tools=",".join(form.getlist("associatedTools")),
             associated_resources=form.get("associatedResources"),
             associated_prompts=form.get("associatedPrompts"),
         )
@@ -227,7 +230,7 @@ async def admin_edit_server(
 
 @admin_router.post("/servers/{server_id}/toggle")
 async def admin_toggle_server(
-    server_id: int,
+    server_id: str,
     request: Request,
     db: Session = Depends(get_db),
     user: str = Depends(require_auth),
@@ -241,7 +244,7 @@ async def admin_toggle_server(
     logs any errors that might occur during the status toggle operation.
 
     Args:
-        server_id (int): The ID of the server whose status to toggle.
+        server_id (str): The ID of the server whose status to toggle.
         request (Request): FastAPI request containing form data with the 'activate' field.
         db (Session): Database session dependency.
         user (str): Authenticated user dependency.
@@ -263,7 +266,7 @@ async def admin_toggle_server(
 
 
 @admin_router.post("/servers/{server_id}/delete")
-async def admin_delete_server(server_id: int, request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
+async def admin_delete_server(server_id: str, request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
     """
     Delete a server via the admin UI.
 
@@ -271,7 +274,7 @@ async def admin_delete_server(server_id: int, request: Request, db: Session = De
     gracefully and logs any errors that occur during the deletion process.
 
     Args:
-        server_id (int): The ID of the server to delete
+        server_id (str): The ID of the server to delete
         request (Request): FastAPI request object (not used but required by route signature).
         db (Session): Database session dependency
         user (str): Authenticated user dependency
@@ -370,7 +373,7 @@ async def admin_list_gateways(
 
 @admin_router.post("/gateways/{gateway_id}/toggle")
 async def admin_toggle_gateway(
-    gateway_id: int,
+    gateway_id: str,
     request: Request,
     db: Session = Depends(get_db),
     user: str = Depends(require_auth),
@@ -383,7 +386,7 @@ async def admin_toggle_gateway(
     determine the new status of the gateway.
 
     Args:
-        gateway_id (int): The ID of the gateway to toggle.
+        gateway_id (str): The ID of the gateway to toggle.
         request (Request): The FastAPI request object containing form data.
         db (Session): The database session dependency.
         user (str): The authenticated user dependency.
@@ -433,12 +436,12 @@ async def admin_ui(
         HTMLResponse: Rendered HTML template for the admin dashboard.
     """
     logger.debug(f"User {user} accessed the admin UI")
-    servers = [server.dict(by_alias=True) for server in await server_service.list_servers(db, include_inactive=include_inactive)]
-    tools = [tool.dict(by_alias=True) for tool in await tool_service.list_tools(db, include_inactive=include_inactive)]
-    resources = [resource.dict(by_alias=True) for resource in await resource_service.list_resources(db, include_inactive=include_inactive)]
-    prompts = [prompt.dict(by_alias=True) for prompt in await prompt_service.list_prompts(db, include_inactive=include_inactive)]
-    gateways = [gateway.dict(by_alias=True) for gateway in await gateway_service.list_gateways(db, include_inactive=include_inactive)]
-    roots = [root.dict(by_alias=True) for root in await root_service.list_roots()]
+    servers = [server.model_dump(by_alias=True) for server in await server_service.list_servers(db, include_inactive=include_inactive)]
+    tools = [tool.model_dump(by_alias=True) for tool in await tool_service.list_tools(db, include_inactive=include_inactive)]
+    resources = [resource.model_dump(by_alias=True) for resource in await resource_service.list_resources(db, include_inactive=include_inactive)]
+    prompts = [prompt.model_dump(by_alias=True) for prompt in await prompt_service.list_prompts(db, include_inactive=include_inactive)]
+    gateways = [gateway.model_dump(by_alias=True) for gateway in await gateway_service.list_gateways(db, include_inactive=include_inactive)]
+    roots = [root.model_dump(by_alias=True) for root in await root_service.list_roots()]
     root_path = settings.app_root_path
     response = request.app.state.templates.TemplateResponse(
         "admin.html",
@@ -452,6 +455,7 @@ async def admin_ui(
             "roots": roots,
             "include_inactive": include_inactive,
             "root_path": root_path,
+            "gateway_tool_name_separator": settings.gateway_tool_name_separator,
         },
     )
 
@@ -486,7 +490,7 @@ async def admin_list_tools(
 
 
 @admin_router.get("/tools/{tool_id}", response_model=ToolRead)
-async def admin_get_tool(tool_id: int, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> ToolRead:
+async def admin_get_tool(tool_id: str, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> ToolRead:
     """
     Retrieve specific tool details for the admin UI.
 
@@ -495,7 +499,7 @@ async def admin_get_tool(tool_id: int, db: Session = Depends(get_db), user: str 
     viewing and management purposes.
 
     Args:
-        tool_id (int): The ID of the tool to retrieve.
+        tool_id (str): The ID of the tool to retrieve.
         db (Session): Database session dependency.
         user (str): Authenticated user dependency.
 
@@ -504,7 +508,7 @@ async def admin_get_tool(tool_id: int, db: Session = Depends(get_db), user: str 
     """
     logger.debug(f"User {user} requested details for tool ID {tool_id}")
     tool = await tool_service.get_tool(db, tool_id)
-    return tool.dict(by_alias=True)
+    return tool.model_dump(by_alias=True)
 
 
 @admin_router.post("/tools/")
@@ -583,7 +587,7 @@ async def admin_add_tool(
 @admin_router.post("/tools/{tool_id}/edit/")
 @admin_router.post("/tools/{tool_id}/edit")
 async def admin_edit_tool(
-    tool_id: int,
+    tool_id: str,
     request: Request,
     db: Session = Depends(get_db),
     user: str = Depends(require_auth),
@@ -611,7 +615,7 @@ async def admin_edit_tool(
     snake-case keys expected by the schemas.
 
     Args:
-        tool_id (int): The ID of the tool to edit.
+        tool_id (str): The ID of the tool to edit.
         request (Request): FastAPI request containing form data.
         db (Session): Database session dependency.
         user (str): Authenticated user dependency.
@@ -639,7 +643,7 @@ async def admin_edit_tool(
         "auth_header_key": form.get("auth_header_key", ""),
         "auth_header_value": form.get("auth_header_value", ""),
     }
-    logger.info(f"Tool update data built: {tool_data}")
+    logger.debug(f"Tool update data built: {tool_data}")
     tool = ToolUpdate(**tool_data)
     try:
         await tool_service.update_tool(db, tool_id, tool)
@@ -653,7 +657,7 @@ async def admin_edit_tool(
 
 
 @admin_router.post("/tools/{tool_id}/delete")
-async def admin_delete_tool(tool_id: int, request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
+async def admin_delete_tool(tool_id: str, request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
     """
     Delete a tool via the admin UI.
 
@@ -662,7 +666,7 @@ async def admin_delete_tool(tool_id: int, request: Request, db: Session = Depend
     and the user must be authenticated to access this route.
 
     Args:
-        tool_id (int): The ID of the tool to delete.
+        tool_id (str): The ID of the tool to delete.
         request (Request): FastAPI request object (not used directly, but required by route signature).
         db (Session): Database session dependency.
         user (str): Authenticated user dependency.
@@ -680,7 +684,7 @@ async def admin_delete_tool(tool_id: int, request: Request, db: Session = Depend
 
 @admin_router.post("/tools/{tool_id}/toggle")
 async def admin_toggle_tool(
-    tool_id: int,
+    tool_id: str,
     request: Request,
     db: Session = Depends(get_db),
     user: str = Depends(require_auth),
@@ -694,7 +698,7 @@ async def admin_toggle_tool(
     logs any errors that might occur during the status toggle operation.
 
     Args:
-        tool_id (int): The ID of the tool whose status to toggle.
+        tool_id (str): The ID of the tool whose status to toggle.
         request (Request): FastAPI request containing form data with the 'activate' field.
         db (Session): Database session dependency.
         user (str): Authenticated user dependency.
@@ -716,7 +720,7 @@ async def admin_toggle_tool(
 
 
 @admin_router.get("/gateways/{gateway_id}", response_model=GatewayRead)
-async def admin_get_gateway(gateway_id: int, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> GatewayRead:
+async def admin_get_gateway(gateway_id: str, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> GatewayRead:
     """Get gateway details for the admin UI.
 
     Args:
@@ -733,7 +737,7 @@ async def admin_get_gateway(gateway_id: int, db: Session = Depends(get_db), user
 
 
 @admin_router.post("/gateways")
-async def admin_add_gateway(request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
+async def admin_add_gateway(request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> JSONResponse:
     """Add a gateway via the admin UI.
 
     Expects form fields:
@@ -755,6 +759,7 @@ async def admin_add_gateway(request: Request, db: Session = Depends(get_db), use
         name=form["name"],
         url=form["url"],
         description=form.get("description"),
+        transport=form.get("transport", "SSE"),
         auth_type=form.get("auth_type", ""),
         auth_username=form.get("auth_username", ""),
         auth_password=form.get("auth_password", ""),
@@ -762,15 +767,26 @@ async def admin_add_gateway(request: Request, db: Session = Depends(get_db), use
         auth_header_key=form.get("auth_header_key", ""),
         auth_header_value=form.get("auth_header_value", ""),
     )
-    await gateway_service.register_gateway(db, gateway)
+    try:
+        await gateway_service.register_gateway(db, gateway)
+        return JSONResponse(
+            content={"message": "Gateway registered successfully!", "success": True},
+            status_code=200,
+        )
 
-    root_path = request.scope.get("root_path", "")
-    return RedirectResponse(f"{root_path}/admin#gateways", status_code=303)
+    except Exception as ex:
+        if isinstance(ex, GatewayConnectionError):
+            return JSONResponse(content={"message": str(ex), "success": False}, status_code=502)
+        if isinstance(ex, ValueError):
+            return JSONResponse(content={"message": str(ex), "success": False}, status_code=400)
+        if isinstance(ex, RuntimeError):
+            return JSONResponse(content={"message": str(ex), "success": False}, status_code=500)
+        return JSONResponse(content={"message": str(ex), "success": False}, status_code=500)
 
 
 @admin_router.post("/gateways/{gateway_id}/edit")
 async def admin_edit_gateway(
-    gateway_id: int,
+    gateway_id: str,
     request: Request,
     db: Session = Depends(get_db),
     user: str = Depends(require_auth),
@@ -797,6 +813,7 @@ async def admin_edit_gateway(
         name=form["name"],
         url=form["url"],
         description=form.get("description"),
+        transport=form.get("transport", "SSE"),
         auth_type=form.get("auth_type", None),
         auth_username=form.get("auth_username", None),
         auth_password=form.get("auth_password", None),
@@ -811,7 +828,7 @@ async def admin_edit_gateway(
 
 
 @admin_router.post("/gateways/{gateway_id}/delete")
-async def admin_delete_gateway(gateway_id: int, request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
+async def admin_delete_gateway(gateway_id: str, request: Request, db: Session = Depends(get_db), user: str = Depends(require_auth)) -> RedirectResponse:
     """
     Delete a gateway via the admin UI.
 
@@ -820,7 +837,7 @@ async def admin_delete_gateway(gateway_id: int, request: Request, db: Session = 
     operation for auditing purposes.
 
     Args:
-        gateway_id (int): The ID of the gateway to delete.
+        gateway_id (str): The ID of the gateway to delete.
         request (Request): FastAPI request object (not used directly but required by the route signature).
         db (Session): Database session dependency.
         user (str): Authenticated user dependency.
