@@ -14,29 +14,36 @@ It handles:
 - Active/inactive gateway management
 """
 
+# Standard
 import asyncio
-import logging
-import uuid
 from datetime import datetime, timezone
+import logging
 from typing import Any, AsyncGenerator, Dict, List, Optional, Set
+import uuid
 
-import httpx
-from filelock import FileLock, Timeout
-from mcp import ClientSession
-from mcp.client.sse import sse_client
-from sqlalchemy import select
-from sqlalchemy.orm import Session
-
+# First-Party
 from mcpgateway.config import settings
 from mcpgateway.db import Gateway as DbGateway
 from mcpgateway.db import SessionLocal
 from mcpgateway.db import Tool as DbTool
 from mcpgateway.schemas import GatewayCreate, GatewayRead, GatewayUpdate, ToolCreate
 from mcpgateway.services.tool_service import ToolService
+from mcpgateway.utils.create_slug import slugify
 from mcpgateway.utils.services_auth import decode_auth
 
+# Third-Party
+from filelock import FileLock, Timeout
+import httpx
+from mcp import ClientSession
+from mcp.client.sse import sse_client
+from mcp.client.streamable_http import streamablehttp_client
+from sqlalchemy import select
+from sqlalchemy.orm import Session
+
 try:
+    # Third-Party
     import redis
+
     REDIS_AVAILABLE = True
 except ImportError:
     REDIS_AVAILABLE = False
@@ -97,7 +104,7 @@ class GatewayService:
         """Initialize the gateway service."""
         self._event_subscribers: List[asyncio.Queue] = []
         self._http_client = httpx.AsyncClient(timeout=settings.federation_timeout, verify=not settings.skip_ssl_verify)
-        self._health_check_interval = GW_HEALTH_CHECK_INTERVAL 
+        self._health_check_interval = GW_HEALTH_CHECK_INTERVAL
         self._health_check_task: Optional[asyncio.Task] = None
         self._active_gateways: Set[str] = set()  # Track active gateway URLs
         self._stream_response = None
@@ -110,9 +117,9 @@ class GatewayService:
 
         if self.redis_url and REDIS_AVAILABLE:
             self._redis_client = redis.from_url(self.redis_url)
-            self._instance_id = str(uuid.uuid4())             # Unique ID for this process
+            self._instance_id = str(uuid.uuid4())  # Unique ID for this process
             self._leader_key = "gateway_service_leader"
-            self._leader_ttl = 40                             # seconds
+            self._leader_ttl = 40  # seconds
         elif settings.cache_type != "none":
             # Fallback: File-based lock
             self._redis_client = None
@@ -122,7 +129,11 @@ class GatewayService:
             self._redis_client = None
 
     async def initialize(self) -> None:
-        """Initialize the service and start health check if this instance is the leader."""
+        """Initialize the service and start health check if this instance is the leader.
+
+        Raises:
+            ConnectionError: When redis ping fails
+        """
         logger.info("Initializing gateway service")
 
         if self._redis_client:
@@ -165,6 +176,7 @@ class GatewayService:
 
         Raises:
             GatewayNameConflictError: If gateway name already exists
+            []: When ExceptionGroup found
         """
         try:
             # Check for name conflicts (both active and inactive)
@@ -180,22 +192,19 @@ class GatewayService:
             auth_type = getattr(gateway, "auth_type", None)
             auth_value = getattr(gateway, "auth_value", {})
 
-            capabilities, tools = await self._initialize_gateway(str(gateway.url), auth_value)
-
-            all_names = [td.name for td in tools]
-
-            existing_tools = db.execute(select(DbTool).where(DbTool.name.in_(all_names))).scalars().all()
-            existing_tool_names = [tool.name for tool in existing_tools]
+            capabilities, tools = await self._initialize_gateway(gateway.url, auth_value, gateway.transport)
 
             tools = [
                 DbTool(
-                    name=tool.name,
-                    url=str(gateway.url),
+                    original_name=tool.name,
+                    original_name_slug=slugify(tool.name),
+                    url=gateway.url,
                     description=tool.description,
                     integration_type=tool.integration_type,
                     request_type=tool.request_type,
                     headers=tool.headers,
                     input_schema=tool.input_schema,
+                    annotations=tool.annotations,
                     jsonpath_filter=tool.jsonpath_filter,
                     auth_type=auth_type,
                     auth_value=auth_value,
@@ -203,20 +212,18 @@ class GatewayService:
                 for tool in tools
             ]
 
-            existing_tools = [tool for tool in tools if tool.name in existing_tool_names]
-            new_tools = [tool for tool in tools if tool.name not in existing_tool_names]
-
             # Create DB model
             db_gateway = DbGateway(
                 name=gateway.name,
-                url=str(gateway.url),
+                slug=slugify(gateway.name),
+                url=gateway.url,
                 description=gateway.description,
+                transport=gateway.transport,
                 capabilities=capabilities,
                 last_seen=datetime.now(timezone.utc),
                 auth_type=auth_type,
                 auth_value=auth_value,
-                tools=new_tools,
-                # federated_tools=existing_tools + new_tools
+                tools=tools,
             )
 
             # Add to DB
@@ -231,19 +238,18 @@ class GatewayService:
             await self._notify_gateway_added(db_gateway)
 
             return GatewayRead.model_validate(gateway)
+        except* GatewayConnectionError as ge:
+            logger.error("GatewayConnectionError in group: %s", ge.exceptions)
+            raise ge.exceptions[0]
         except* ValueError as ve:
             logger.error("ValueErrors in group: %s", ve.exceptions)
+            raise ve.exceptions[0]
         except* RuntimeError as re:
             logger.error("RuntimeErrors in group: %s", re.exceptions)
+            raise re.exceptions[0]
         except* BaseException as other:  # catches every other sub-exception
             logger.error("Other grouped errors: %s", other.exceptions)
-        # except IntegrityError as ex:
-        #     logger.error(f"Error adding gateway: {ex}")
-        #     db.rollback()
-        #     raise GatewayError(f"Gateway already exists: {gateway.name}")
-        # except Exception as e:
-        #     db.rollback()
-        #     raise GatewayError(f"Failed to register gateway: {str(e)}")
+            raise other.exceptions[0]
 
     async def list_gateways(self, db: Session, include_inactive: bool = False) -> List[GatewayRead]:
         """List all registered gateways.
@@ -263,7 +269,7 @@ class GatewayService:
         gateways = db.execute(query).scalars().all()
         return [GatewayRead.model_validate(g) for g in gateways]
 
-    async def update_gateway(self, db: Session, gateway_id: int, gateway_update: GatewayUpdate) -> GatewayRead:
+    async def update_gateway(self, db: Session, gateway_id: str, gateway_update: GatewayUpdate) -> GatewayRead:
         """Update a gateway.
 
         Args:
@@ -302,10 +308,13 @@ class GatewayService:
             # Update fields if provided
             if gateway_update.name is not None:
                 gateway.name = gateway_update.name
+                gateway.slug = slugify(gateway_update.name)
             if gateway_update.url is not None:
-                gateway.url = str(gateway_update.url)
+                gateway.url = gateway_update.url
             if gateway_update.description is not None:
                 gateway.description = gateway_update.description
+            if gateway_update.transport is not None:
+                gateway.transport = gateway_update.transport
 
             if getattr(gateway, "auth_type", None) is not None:
                 gateway.auth_type = gateway_update.auth_type
@@ -317,9 +326,31 @@ class GatewayService:
             # Try to reinitialize connection if URL changed
             if gateway_update.url is not None:
                 try:
-                    capabilities, _ = await self._initialize_gateway(gateway.url, gateway.auth_value)
+                    capabilities, tools = await self._initialize_gateway(gateway.url, gateway.auth_value, gateway.transport)
+                    new_tool_names = [tool.name for tool in tools]
+
+                    for tool in tools:
+                        existing_tool = db.execute(select(DbTool).where(DbTool.original_name == tool.name).where(DbTool.gateway_id == gateway_id)).scalar_one_or_none()
+                        if not existing_tool:
+                            gateway.tools.append(
+                                DbTool(
+                                    original_name=tool.name,
+                                    original_name_slug=slugify(tool.name),
+                                    url=gateway.url,
+                                    description=tool.description,
+                                    integration_type=tool.integration_type,
+                                    request_type=tool.request_type,
+                                    headers=tool.headers,
+                                    input_schema=tool.input_schema,
+                                    jsonpath_filter=tool.jsonpath_filter,
+                                    auth_type=gateway.auth_type,
+                                    auth_value=gateway.auth_value,
+                                )
+                            )
+
                     gateway.capabilities = capabilities
-                    gateway.last_seen = datetime.utcnow()
+                    gateway.tools = [tool for tool in gateway.tools if tool.original_name in new_tool_names]  # keep only still-valid rows
+                    gateway.last_seen = datetime.now(timezone.utc)
 
                     # Update tracking with new URL
                     self._active_gateways.discard(gateway.url)
@@ -327,7 +358,7 @@ class GatewayService:
                 except Exception as e:
                     logger.warning(f"Failed to initialize updated gateway: {e}")
 
-            gateway.updated_at = datetime.utcnow()
+            gateway.updated_at = datetime.now(timezone.utc)
             db.commit()
             db.refresh(gateway)
 
@@ -341,7 +372,7 @@ class GatewayService:
             db.rollback()
             raise GatewayError(f"Failed to update gateway: {str(e)}")
 
-    async def get_gateway(self, db: Session, gateway_id: int, include_inactive: bool = False) -> GatewayRead:
+    async def get_gateway(self, db: Session, gateway_id: str, include_inactive: bool = False) -> GatewayRead:
         """Get a specific gateway by ID.
 
         Args:
@@ -364,7 +395,7 @@ class GatewayService:
 
         return GatewayRead.model_validate(gateway)
 
-    async def toggle_gateway_status(self, db: Session, gateway_id: int, activate: bool) -> GatewayRead:
+    async def toggle_gateway_status(self, db: Session, gateway_id: str, activate: bool) -> GatewayRead:
         """Toggle gateway active status.
 
         Args:
@@ -387,16 +418,39 @@ class GatewayService:
             # Update status if it's different
             if gateway.is_active != activate:
                 gateway.is_active = activate
-                gateway.updated_at = datetime.utcnow()
+                gateway.updated_at = datetime.now(timezone.utc)
 
                 # Update tracking
                 if activate:
                     self._active_gateways.add(gateway.url)
                     # Try to initialize if activating
                     try:
-                        capabilities, tools = await self._initialize_gateway(gateway.url, gateway.auth_value)
-                        gateway.capabilities = capabilities.dict()
-                        gateway.last_seen = datetime.utcnow()
+                        capabilities, tools = await self._initialize_gateway(gateway.url, gateway.auth_value, gateway.transport)
+                        new_tool_names = [tool.name for tool in tools]
+
+                        for tool in tools:
+                            existing_tool = db.execute(select(DbTool).where(DbTool.original_name == tool.name).where(DbTool.gateway_id == gateway_id)).scalar_one_or_none()
+                            if not existing_tool:
+                                gateway.tools.append(
+                                    DbTool(
+                                        original_name=tool.name,
+                                        original_name_slug=slugify(tool.name),
+                                        url=gateway.url,
+                                        description=tool.description,
+                                        integration_type=tool.integration_type,
+                                        request_type=tool.request_type,
+                                        headers=tool.headers,
+                                        input_schema=tool.input_schema,
+                                        jsonpath_filter=tool.jsonpath_filter,
+                                        auth_type=gateway.auth_type,
+                                        auth_value=gateway.auth_value,
+                                    )
+                                )
+
+                        gateway.capabilities = capabilities
+                        gateway.tools = [tool for tool in gateway.tools if tool.original_name in new_tool_names]  # keep only still-valid rows
+
+                        gateway.last_seen = datetime.now(timezone.utc)
                     except Exception as e:
                         logger.warning(f"Failed to initialize reactivated gateway: {e}")
                 else:
@@ -439,11 +493,11 @@ class GatewayService:
                 "description": gateway.description,
                 "is_active": gateway.is_active,
             },
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
         }
         await self._publish_event(event)
 
-    async def delete_gateway(self, db: Session, gateway_id: int) -> None:
+    async def delete_gateway(self, db: Session, gateway_id: str) -> None:
         """Permanently delete a gateway.
 
         Args:
@@ -509,7 +563,7 @@ class GatewayService:
             result = response.json()
 
             # Update last seen timestamp
-            gateway.last_seen = datetime.utcnow()
+            gateway.last_seen = datetime.now(timezone.utc)
 
             if "error" in result:
                 raise GatewayError(f"Gateway error: {result['error'].get('message')}")
@@ -566,16 +620,23 @@ class GatewayService:
                     headers = decode_auth(auth_data)
 
                     # Perform the GET and raise on 4xx/5xx
-                    async with client.stream("GET", gateway.url, headers=headers) as response:
-                        # This will raise immediately if status is 4xx/5xx
-                        response.raise_for_status()
+                    if (gateway.transport).lower() == "sse":
+                        timeout = httpx.Timeout(settings.health_check_timeout)
+                        async with client.stream("GET", gateway.url, headers=headers, timeout=timeout) as response:
+                            # This will raise immediately if status is 4xx/5xx
+                            response.raise_for_status()
+                    elif (gateway.transport).lower() == "streamablehttp":
+                        async with streamablehttp_client(url=gateway.url, headers=headers, timeout=settings.health_check_timeout) as (read_stream, write_stream, _get_session_id):
+                            async with ClientSession(read_stream, write_stream) as session:
+                                # Initialize the session
+                                response = await session.initialize()
 
                     # Mark successful check
-                    gateway.last_seen = datetime.utcnow()
+                    gateway.last_seen = datetime.now(timezone.utc)
 
                 except Exception:
                     await self._handle_gateway_failure(gateway)
-                    
+
         # All gateways passed
         return True
 
@@ -624,12 +685,13 @@ class GatewayService:
         finally:
             self._event_subscribers.remove(queue)
 
-    async def _initialize_gateway(self, url: str, authentication: Optional[Dict[str, str]] = None) -> Any:
+    async def _initialize_gateway(self, url: str, authentication: Optional[Dict[str, str]] = None, transport: str = "SSE") -> Any:
         """Initialize connection to a gateway and retrieve its capabilities.
 
         Args:
             url: Gateway URL
             authentication: Optional authentication headers
+            transport: Transport type ("SSE" or "StreamableHTTP")
 
         Returns:
             Capabilities dictionary as provided by the gateway.
@@ -671,7 +733,47 @@ class GatewayService:
 
                 return capabilities, tools
 
-            capabilities, tools = await connect_to_sse_server(url, authentication)
+            async def connect_to_streamablehttp_server(server_url: str, authentication: Optional[Dict[str, str]] = None):
+                """
+                Connect to an MCP server running with Streamable HTTP transport
+
+                Args:
+                    server_url: URL to connect to the server
+                    authentication: Authentication headers for connection to URL
+
+                Returns:
+                    list, list: List of capabilities and tools
+                """
+                if authentication is None:
+                    authentication = {}
+                # Store the context managers so they stay alive
+                decoded_auth = decode_auth(authentication)
+
+                # Use async with for both streamablehttp_client and ClientSession
+                async with streamablehttp_client(url=server_url, headers=decoded_auth) as (read_stream, write_stream, _get_session_id):
+                    async with ClientSession(read_stream, write_stream) as session:
+                        # Initialize the session
+                        response = await session.initialize()
+                        # if get_session_id:
+                        #     session_id = get_session_id()
+                        #     if session_id:
+                        #         print(f"Session ID: {session_id}")
+                        capabilities = response.capabilities.model_dump(by_alias=True, exclude_none=True)
+                        response = await session.list_tools()
+                        tools = response.tools
+                        tools = [tool.model_dump(by_alias=True, exclude_none=True) for tool in tools]
+                        tools = [ToolCreate.model_validate(tool) for tool in tools]
+                        for tool in tools:
+                            tool.request_type = "STREAMABLEHTTP"
+
+                return capabilities, tools
+
+            capabilities = {}
+            tools = []
+            if transport.lower() == "sse":
+                capabilities, tools = await connect_to_sse_server(url, authentication)
+            elif transport.lower() == "streamablehttp":
+                capabilities, tools = await connect_to_streamablehttp_server(url, authentication)
 
             return capabilities, tools
         except Exception as e:
@@ -718,7 +820,7 @@ class GatewayService:
                         logger.error(f"Health check run failed: {str(e)}")
 
                     await asyncio.sleep(self._health_check_interval)
-                
+
                 else:
                     # FileLock-based leader fallback
                     try:
@@ -776,7 +878,7 @@ class GatewayService:
                 "description": gateway.description,
                 "is_active": gateway.is_active,
             },
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
         }
         await self._publish_event(event)
 
@@ -795,7 +897,7 @@ class GatewayService:
                 "url": gateway.url,
                 "is_active": True,
             },
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
         }
         await self._publish_event(event)
 
@@ -814,7 +916,7 @@ class GatewayService:
                 "url": gateway.url,
                 "is_active": False,
             },
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
         }
         await self._publish_event(event)
 
@@ -828,7 +930,7 @@ class GatewayService:
         event = {
             "type": "gateway_deleted",
             "data": gateway_info,
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
         }
         await self._publish_event(event)
 
@@ -842,7 +944,7 @@ class GatewayService:
         event = {
             "type": "gateway_removed",
             "data": {"id": gateway.id, "name": gateway.name, "is_active": False},
-            "timestamp": datetime.utcnow().isoformat(),
+            "timestamp": datetime.now(timezone.utc).isoformat(),
         }
         await self._publish_event(event)
 
