@@ -10,7 +10,7 @@ Authors: Mihai Criveti
 __author__ = "Mihai Criveti"
 __copyright__ = "Copyright 2025"
 __license__ = "Apache 2.0"
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 __description__ = "IBM Consulting Assistants - Extensions API Library"
 __url__ = "https://ibm.github.io/mcp-context-forge/"
 __download_url__ = "https://github.com/IBM/mcp-context-forge"
