@@ -9,8 +9,8 @@ This module defines configuration settings for the MCP Gateway using Pydantic.
 It loads configuration from environment variables with sensible defaults.
 
 Environment variables:
-- APP_NAME: Gateway name (default: "MCP Gateway")
-- HOST: Host to bind to (default: "0.0.0.0")
+- APP_NAME: Gateway name (default: "MCP_Gateway")
+- HOST: Host to bind to (default: "127.0.0.1")
 - PORT: Port to listen on (default: 4444)
 - DATABASE_URL: SQLite database URL (default: "sqlite:///./mcp.db")
 - BASIC_AUTH_USER: Admin username (default: "admin")
@@ -29,14 +29,16 @@ Environment variables:
 - HEALTH_CHECK_INTERVAL: Gateway health check interval (default: 60)
 """
 
-import json
+# Standard
 from functools import lru_cache
 from importlib.resources import files
+import json
 from pathlib import Path
 from typing import Annotated, Any, Dict, List, Optional, Set, Union
 
-import jq
+# Third-Party
 from fastapi import HTTPException
+import jq
 from jsonpath_ng.ext import parse
 from jsonpath_ng.jsonpath import JSONPath
 from pydantic import Field, field_validator
@@ -47,7 +49,7 @@ class Settings(BaseSettings):
     """MCP Gateway configuration settings."""
 
     # Basic Settings
-    app_name: str = Field("MCP Gateway", env="APP_NAME")
+    app_name: str = Field("MCP_Gateway", env="APP_NAME")
     host: str = Field("127.0.0.1", env="HOST")
     port: int = Field(4444, env="PORT")
     database_url: str = "sqlite:///./mcp.db"
@@ -191,12 +193,18 @@ class Settings(BaseSettings):
     session_ttl: int = 3600
     message_ttl: int = 600
 
+    # streamable http transport
+    use_stateful_sessions: bool = False  # Set to False to use stateless sessions without event store
+    json_response_enabled: bool = True  # Enable JSON responses instead of SSE streams
+
     # Development
     dev_mode: bool = False
     reload: bool = False
     debug: bool = False
 
     model_config = SettingsConfigDict(env_file=".env", env_file_encoding="utf-8", case_sensitive=False, extra="ignore")
+
+    gateway_tool_name_separator: str = "-"
 
     @property
     def api_key(self) -> str:
