@@ -8,15 +8,18 @@ Authors: Mihai Criveti
 Tests for the MCP Gateway SSE transport implementation.
 """
 
+# Standard
 import asyncio
 import json
 from unittest.mock import Mock
 
-import pytest
-from fastapi import Request
-from sse_starlette.sse import EventSourceResponse
-
+# First-Party
 from mcpgateway.transports.sse_transport import SSETransport
+
+# Third-Party
+from fastapi import Request
+import pytest
+from sse_starlette.sse import EventSourceResponse
 
 
 @pytest.fixture
